+++ src/EditorObjectLayering.hpp
@@ -1,0 +1,7 @@
+#include <Geode/Geode.hpp>
+
+using namespace geode::prelude;
+
+namespace EditorObjectLayering {
+    void updateObjectLayering(GameObject*);
+}