+++ src/SetGroupIDLayer.cpp
@@ -5,26 +5,79 @@ using namespace geode::prelude;
 #include <Geode/modify/SetGroupIDLayer.hpp>
 #include "utils.hpp"
 
-class $modify(PolzSetGroupIDLayer, SetGroupIDLayer) {
-    void onCurrentEditorGroup(CCObject*) {
-        auto& m_targetObject = from<GameObject*>(this, 0x1bc);
-        auto& m_targetObjects = from<CCArray*>(this, 0x1c0);
-        auto& m_groupLabel = from<CCLabelBMFont*>(this, 0x1c4);
+class ObjectGroupInput : public CCLayer, TextInputDelegate {
+public:
+    SetGroupIDLayer* m_parent;
+    CCTextInputNode* m_input;
 
+    static ObjectGroupInput* create(SetGroupIDLayer* parent) {
+        ObjectGroupInput* pRet = new ObjectGroupInput();
+        if (pRet && pRet->init(parent)) {
+            pRet->autorelease();
+            return pRet;
+        }
+        CC_SAFE_DELETE(pRet);
+        return nullptr;
+    }
+
+    bool init(SetGroupIDLayer* parent) {
+        if (!CCLayer::init()) return false;
+        this->m_parent = parent;
+
+        this->m_input = CCTextInputNode::create(60.f, 35.f, "G", "bigFont.fnt");
+        m_input->m_delegate = this;
+        m_input->setLabelPlaceholderColor({120,120,120});
+        m_input->setAllowedChars("0123456789");
+        m_input->m_maxLabelLength = 3;
+        m_input->setString(parent->m_groupLabel->getString());
+
+        auto bg = extension::CCScale9Sprite::create("square02_small.png");
+        bg->setContentSize({60.f, 35.f});
+        bg->setOpacity(75);
+
+        addChild(m_input);
+        addChild(bg, -1);
+
+        return true;
+    }
+
+    virtual void textChanged(CCTextInputNode* input) override {
+        log::debug("Current Group: {}", std::atoi(input->getString().c_str()));
+        auto value = std::atoi(input->getString().c_str());
+
+        if (m_parent->m_targetObject) {
+            m_parent->m_targetObject->m_editorLayer = value;
+        }
+        else if (m_parent->m_targetObjects) {
+            for (auto obj : CCArrayExt<GameObject*>(m_parent->m_targetObjects)) {
+                obj->m_editorLayer = value;
+            }
+        }
+
+        m_parent->m_groupLabel->setString(std::to_string(value).c_str());
+    }
+};
+
+class $modify(PolzSetGroupIDLayer, SetGroupIDLayer) {
+    struct Fields {
+        ObjectGroupInput* m_objectGroupInput;
+    };
+
+    void onCurrentEditorGroup(CCObject*) {
         LevelEditorLayer* lel = typeinfo_cast<LevelEditorLayer*>(CCDirector::sharedDirector()->getRunningScene()->getChildren()->objectAtIndex(0));
 
         if (lel->m_groupIDFilter == -1) return;
 
-        if (m_targetObject) {
-            m_targetObject->m_editorLayer = lel->m_groupIDFilter;
+        if (this->m_targetObject) {
+            this->m_targetObject->m_editorLayer = lel->m_groupIDFilter;
         }
-        else if (m_targetObjects) {
-            for (auto obj : CCArrayExt<GameObject*>(m_targetObjects)) {
+        else if (this->m_targetObjects) {
+            for (auto obj : CCArrayExt<GameObject*>(this->m_targetObjects)) {
                 obj->m_editorLayer = lel->m_groupIDFilter;
             }
         }
 
-        m_groupLabel->setString(std::to_string(lel->m_groupIDFilter).c_str());
+        this->m_groupLabel->setString(std::to_string(lel->m_groupIDFilter).c_str());
     }
 
     bool init(GameObject* p0, CCArray* p1) {
@@ -39,6 +92,16 @@ class $modify(PolzSetGroupIDLayer, SetGroupIDLayer) {
         onCurrentGroup->setID("onCurrentGroup-button"_spr);
         this->m_buttonMenu->addChild(onCurrentGroup);
 
+        this->m_fields->m_objectGroupInput = ObjectGroupInput::create(this);
+        this->m_mainLayer->addChild(this->m_fields->m_objectGroupInput, 5);
+        this->m_fields->m_objectGroupInput->setPosition(this->m_groupLabel->getPosition());
+        this->m_groupLabel->setVisible(false);
+
         return true;
+    }
+
+    void updateGroupID() {
+        SetGroupIDLayer::updateGroupID();
+        this->m_fields->m_objectGroupInput->m_input->setString(this->m_groupLabel->getString());
     }
 };