+++ changelog.md
@@ -1,8 +1,25 @@
+# 1.0.2
+- Added Blending Indicators.
+- Added Object Info.
+- Added Custom Fade Time Input.
+- Added Editor Freeze Fix.
+- Added Grid Size Controls.
+- Added Custom Object Group Input.
+- Added Zoom Bypass.
+- Added Experimental Object Layering.
+- Added Editor Unpausing on ESC.
+- Added Editor Extension.
+- Fixed Custom Rotate Buttons.
+
 # 1.0.1
-- Added Custom Move Buttons (1/2, 1/4, 1/8, Unit).
+- Added Mod Settings.
+- Added Custom Move Buttons (1/2, 1/4, 1/8, 0.5, 0.1).
 - Added Set Current Group button.
-- Added Select Filter toggler.
 - Added Start Flipped toggler.
+- Added Scroll Zoom.
+- Added Move Object crash fix (nice RobTop).
+- Added Hide Editor UI button.
+- Added Duration Lines.
 
 # 1.0.0
 - Added ability to color objects to 3DL and White.